# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a game management background. you can use the api key `ApiKeyAuth` to test the authorization filters.",
        "title": "gin-gorm-admin API",
        "termsOfService": "https://github.com",
        "contact": {
            "name": "conjurer",
            "url": "https:/github.com/dot123",
            "email": "conjurer888888@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
        "/monitor/index": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器状态"
                ],
                "summary": "获取服务器状态",
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/monitor.Server"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/msg/notice": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公告管理"
                ],
                "summary": "获取公告列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.NoticeGetPageResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公告管理"
                ],
                "summary": "更新公告",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.NoticeUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公告管理"
                ],
                "summary": "新建公告",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.NoticeCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/msg/notice/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公告管理"
                ],
                "summary": "删除公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/public/uploadFile": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/user": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "修改用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "新建用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UserCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "获取用户信息",
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.UserInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.UserGetPageReqResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseData"
                        }
                    },
                    "500": {
                        "description": "失败结果",
                        "schema": {
                            "$ref": "#/definitions/ginx.ResponseFail"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ginx.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "ginx.ResponseFail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "monitor.Cpu": {
            "type": "object",
            "properties": {
                "cores": {
                    "type": "integer"
                },
                "cpus": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "monitor.Disk": {
            "type": "object",
            "properties": {
                "totalGb": {
                    "type": "integer"
                },
                "totalMb": {
                    "type": "integer"
                },
                "usedGb": {
                    "type": "integer"
                },
                "usedMb": {
                    "type": "integer"
                },
                "usedPercent": {
                    "type": "integer"
                }
            }
        },
        "monitor.Os": {
            "type": "object",
            "properties": {
                "compiler": {
                    "type": "string"
                },
                "goVersion": {
                    "type": "string"
                },
                "goos": {
                    "type": "string"
                },
                "numCpu": {
                    "type": "integer"
                },
                "numGoroutine": {
                    "type": "integer"
                }
            }
        },
        "monitor.Rrm": {
            "type": "object",
            "properties": {
                "totalMb": {
                    "type": "integer"
                },
                "usedMb": {
                    "type": "integer"
                },
                "usedPercent": {
                    "type": "integer"
                }
            }
        },
        "monitor.Server": {
            "type": "object",
            "properties": {
                "cpu": {
                    "$ref": "#/definitions/monitor.Cpu"
                },
                "disk": {
                    "$ref": "#/definitions/monitor.Disk"
                },
                "os": {
                    "$ref": "#/definitions/monitor.Os"
                },
                "ram": {
                    "$ref": "#/definitions/monitor.Rrm"
                }
            }
        },
        "schema.Notice": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schema.NoticeCreateReq": {
            "type": "object",
            "required": [
                "content",
                "endTime",
                "operator",
                "startTime",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 2
                },
                "endTime": {
                    "type": "string"
                },
                "operator": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "startTime": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                }
            }
        },
        "schema.NoticeGetPageResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Notice"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "schema.NoticeUpdateReq": {
            "type": "object",
            "required": [
                "content",
                "endTime",
                "id",
                "operator",
                "startTime",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 2
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "startTime": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                }
            }
        },
        "schema.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "schema.UserCreateReq": {
            "type": "object",
            "required": [
                "password",
                "user_type",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 6
                },
                "user_type": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "schema.UserGetPageReqResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.User"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "schema.UserInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.UserUpdateReq": {
            "type": "object",
            "required": [
                "id",
                "user_type",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "maxLength": 12
                },
                "user_type": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}